+++ docs/src/markdown/about/changelog.md
@@ -7,6 +7,8 @@
     RGB space.
 -   **FIX**: When a `Color` instance is passed as an input to the `Color` constructor, use a more reliable check to
     ensure the color instance is of a compatible color space.
+-   **FIX**: Fix regression where that caused `longer` hue interpolation to no longer work with `discrete`
+    interpolation.
 
 ## 4.7.1